Handle apps with zero or multiple certificates. SALM-76 and SALM-70.
This change addresses SALM-76. The converter was always providing
exactly one application certificate in the configuration, regardless
of how many certificates were actually requested in the conversion
request.

The fix here is to include a Vec of cert configs in the EnclaveConfig
instead of just a single cert config. We also need to change the code
that acquires app certs to get one cert per configured cert.

This trivially also fixes SALM-70, since if there are no configured
certificates, we'll skip trying to get an app cert.