Quick hack for ZIRC-73. Foreground the proxy process.
I think the problem in ZIRC-73 is that the nitro-cli command
is exiting after starting the enclave, which causes the parent
container to exit. The quick fix here is to foreground the
proxy process in the parent, so the shell script won't exit
until the proxy exits. I think this may lead to the parent
container not exiting when the enclave application process
exiting, but at least we should be able to run some applications.