Merged in mihir/client-gen-refactoring (pull request #4364)
client-generator code clean up & refactor for future changes

- move client_generator_base to webservice-client
- move codegen attr filtering to api_model_parser:
Previously filtering was done by a separate method named .filtered()
which would remove all fields containing `codegen(skip)` or were empty
attributes like #[derive()]. But this is the job of api_model_parser
and this will also help remove CodegenAttribute from
client-generator-model.
- use TypeMetaData instead of CodegenAttribute:
`#[codegen(...)]` are interpreted by api_model_parser for
`skip` and `custom_type` and there is a type named `CodegenAnnotation`
representing it. These are no more meant for generator crate.
So, this commit completely removed codegen from the view of generator.
Generator crate was also using CodegenAttribute for indicating usage
as query params and inp/out type. The same is now being done by
`TypeMetaData`. There is a meta_data field in TypeDef for these purposes
only. There one can associate data with the TypeDef via code in
parser/generator.
- make client_generation in models feature gated:
Various model crates contain `client_generation.rs` which
is only needed by client-generator crates. So, made those modules
in these crates feature gated to optimize the build times
of other crates.
- move parse_api_model from trait to separate impl
Before this commit, we had parse_api_model() as a method
of the ParsableApiModel trait.
ParsableApiModel is only meant to describe how to parse a model,
but methods like `api_model_name()` and `parse_api_model()` were in it.
This is not intuitive and also reduces flexibility if we want to use
separate configuration for different parsers.
So, this commit adds a new type ApiModelParser which is esentially
contains configuration for the current parse and using that we call
parse_api_model().
- enforce custom type if custom serialize/deserialize
A custom type is something for which stubs can't be generated
and you need to implement it yourself in each client.
We want to force the type to be a custom type if it doesn't
derive both Serialize and Deserialize. If it derives one of those,
then we don't enfore anything assuming the other one has to be
inverse of the derived one and we generally have a custom Deserialize
impl only for backcompat which isn't required in generated code.
Also, if any of its field/variant has "both" `serialize_with`
and `deserialize_with`, we require type to be custom for the
same reason.
This should not be applicable to query params
as they anyway have a custom serialize/deserialize
impl.
- remove custom Deserialize from FpeOptions
In this PR, a check was added to force types with both
custom serialize and deserialize to be custom type.
FpeOptions had both and the custom Deserialize doesn't seem
necessary as we do the same with untagged enums at other places
also.
- Make PluginVersion & GetAllResponse custom types.
- add builder methods like deny_missing_docs, etc to ApiModelParser.
- Translate Identifier to String for time being.

Approved-by: Pieter Agten