Merged in SALM-245 (pull request #5321)
Fetch enclave kernel before building converter container and not in ci.sh. SALM-245.

The enclave kernel does not need to be installed in the runtime environment, it is sufficient if it is present in the converter container since the converted image contains the enclave kernel.

Note that this change will increase the size of the converted image since the kernel is included in the eif file.

Approved-by: daniel@fortanix.com