Merged in aj/SALM-328 (pull request #5949)
SALM-328 Update ownership of std stream fs of the current process in the start up binary before launching the client program 

Update the stdin/out/err file descriptors of the current process to be owned by the user/group of the client process. stdout of the process eventually points to /dev/console so the above issue is fixed. We also update the ownership of stdin/stderr to avoid similar issues in the future where the client program might be running as non-root user and redirecting it’s stdin/out/err to a different location.


Approved-by: daniel@fortanix.com