Merged in aj/SALM-345 (pull request #6078)
SALM-345 Link enclave startup with musl rather than glibc

While running some applications in the nitro enclave, we see segfaults with the enclave startup program when it verifies and sets up the user/group of the client program. This occurred because glibc doesn’t support linking with libnss statically and always links dynamically. To avoid this problem, link with musl.


Approved-by: Nikita Shyrei
Approved-by: daniel@fortanix.com