Merged in aj/SALM-350 (pull request #6100)
SALM-350 Pass input image environment variables in the enclave manifest instead of setting it during the enclave image build process.

This is a fix for conversion failures seen in images where the environment variable values contain spaces or special characters. They cause the enclave image build to break due to absence of shell escaping in the input+enclave image Dockerfile.

Approved-by: daniel@fortanix.com