Merged in aj/SALM-371 (pull request #6146)
SALM-371 Update enclave base image which includes trusted CA certs needed to access DSM

Add functionality to fetch sdkms version and print to console so that we know we are connected to the sdkms server. Right now, if the endpoint or api key is incorrect or if the client cant connect to sdkms the app hangs.

Approved-by: Nikita Shyrei
Approved-by: daniel@fortanix.com