Merged in aj/SALM-388 (pull request #6287)
Check disk space before creating blockfile 

Recently multiple conversions in the nitro converter were failing because the converter could not create the read-only blockfile. This was occurring because the converter was running out of memory. Add a check in the configure script to check disk space available before creating blockfile.

Approved-by: daniel@fortanix.com