Update everything in roche to use sgx_pkix version 1.4.
This updates everything that has a direct dependency
on sgx_pkix to sgx_pkix version 1.4. I eventually
need this in senclave and sdkms-cli to pick up a new
OID definition. I decided to update everything to keep
all of the software on the latest version. This is a
relatively safe crate to update since the changes
are fairly minor on the 0.1.x branch.

No new unit tests are required for this change because
it's just updating crate dependencies.