Merged in aj/SALM-334-dev-branch (pull request #6661)
SALM-334 Provide a converter option to enable usage of DSM keys for FS persistance

Add an optional parameter enable_overlay_filesystem_persistence of boolean type to the converter API. When set to true, it will allow the salmiac app to use DSM keys for overlayfs encryption. By default it is set to false.

Document some environment variables used by salmiac. This list is not exhaustive and can be worked on in a different task.

Update zircon submodule to test the changes made to the converter API via app tests.


Approved-by: John Foley
Approved-by: Denis Legalov