Merged in aj/SALM-433 (pull request #6782)
SALM-433 Null terminate the app cert while connecting to DSM 

The appcert was not null terminated before it was used for connecting to DSM, this should’ve ideally returned an error and caused the salmiac runner to exit instead of hang. The fix in the PR ensures that we null terminate the appcert.

The unit test was reading the appcert out of a file and appending a 0 to its contents so this issue was never caught in CI. Modified the unit test to not null terminate the appcert while reading it from file. 

Approved-by: John Foley
Approved-by: Dean Menezes