[vsock-proxy parent] Don't have separate lists for device settings and capture devices
In the parent packet proxy setup code, we used to have two separate lists for capture
devices (for our PCAP setup) and device settings (for communicating to the enclave).
Some code assumed that these two lists were in a particular matching order, but this
invariant wasn't upheld everywhere.

Instead, refactor the code to not carry the captures devices but just rely on device
name everywhere. This means we only need a single list which avoids problems in
general.